+++ mex/editor/rules/transform.py
@@ -101,6 +101,18 @@ def _transform_model_to_input_config(  # noqa: PLR0911
     editable: bool,  # noqa: FBT001
 ) -> InputConfig:
     """Determine the input type for a given field of a given model."""
+    if field_name == "identifierInPrimarySource":
+        return InputConfig(
+            editable_text=False,
+            editable_badge=False,
+            editable_identifier=False,
+            editable_href=False,
+            allow_additive=False,
+            render_textarea=False,
+            allow_preventive=False,
+            allow_subtractive=False,
+        )
+
     if field_name in REFERENCE_FIELDS_BY_CLASS_NAME[entity_type]:
         return InputConfig(
             editable_identifier=editable,
@@ -156,6 +168,7 @@ def _transform_model_to_input_config(  # noqa: PLR0911
             editable_text=editable,
             allow_additive=editable,
         )
+
     return InputConfig()
 
 
@@ -192,6 +205,7 @@ def _transform_model_to_editor_primary_sources(
     """With a model and rules, attach an editor primary source to the field."""
     primary_source_id = _get_primary_source_id_from_model(model)
     primary_source_name = transform_value(primary_source_id)
+
     for field_name in type(model).model_fields:
         if field_name in fields_by_name:
             editor_values = _transform_model_values_to_editor_values(
@@ -253,7 +267,13 @@ def transform_models_to_fields(
         )
         for (field_name, entity_type) in mergeable_fields
     }
-
+    if extracted_items:
+        fields_by_name["identifierInPrimarySource"] = EditorField(
+            name="identifierInPrimarySource",
+            value_type=["str"],
+            primary_sources=[],
+            is_required=False,
+        )
     for extracted in extracted_items:
         _transform_model_to_editor_primary_sources(
             fields_by_name,